pyproject.toml
[tool.poetry]
name = "docudex"
version = "0.1.0"
description = "DocuDex is a CLI tool for efficient document interaction and analysis."
authors = ["Lohith Srinivasaiah <lohith.srinivasaiah+github@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
homepage = "https://github.com/lohithsrinivasaiah/docudex"
repository = "https://github.com/lohithsrinivasaiah/docudex"
documentation = "https://github.com/lohithsrinivasaiah/docudex#readme"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
toml = "^0.10.2"
ruamel-yaml = "^0.18.6"
rich = "^13.7.1"
urwid = "^2.6.15"

[tool.poetry.scripts]
docudex = "docudex.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
